requirements.txt
streamlit==1.32.0
openai==1.14.0
anthropic==0.20.0
langchain==0.1.12
langchain-community==0.0.28
langchain-openai==0.0.8
faiss-cpu==1.7.4
sentence-transformers==2.5.1
chromadb==0.4.24
Pillow==10.2.0
pytesseract==0.3.10
easyocr==1.7.1
pydantic==2.6.3
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
sympy==1.12
scipy==1.12.0
tiktoken==0.6.0
httpx==0.27.0
aiofiles==23.2.1
tqdm==4.66.2
requests==2.31.0
openai-whisper==20231117
